Finalizing the components according to the review
